requirements.txt
tensorflow
pillow
matplotlib
numpy
seaborn
ipykernel
notebook